+++ frontend/app/src/screens/objects/relationships/ui/relationship-combobox-list.tsx
@@ -2,7 +2,7 @@ import { ComboboxEmpty, ComboboxItem, ComboboxList } from "@/components/ui/combo
 import { Spinner } from "@/components/ui/spinner";
 import { useSchema } from "@/hooks/useSchema";
 import ErrorScreen from "@/screens/errors/error-screen";
-import { relationshipsInfiniteQueryOptions } from "@/screens/objects/relationships/domain/get-relationships";
+import { relationshipsInfiniteQueryOptions } from "@/screens/objects/relationships/domain/get-relationships/get-relationships.query";
 import { RelationshipNode } from "@/screens/objects/relationships/domain/types";
 import { AddRelationshipAction } from "@/screens/objects/relationships/ui/add-relationship-action";
 import { debounce } from "@/utils/common";
